requirements.txt
yt-dlp
openai-whisper
ffmpeg
feedparser
pandas
google-generativeai
kaggle 